requirements.txt
torch
torchvision
torchaudio
opencv-python
matplotlib
numpy
pandas
tqdm
ipython
Pillow
scipy